Change `Status.__repr__` to default `__repr__`, to conform to usual expectations.
Before this change, `__repr__` is identical to `__str__`. There are countless articles on the web explaining the expected difference, for example:

* https://www.tutorialspoint.com/str-vs-repr-in-python

Quoting "Key differences":

* str(): Make object readable; Generate output to end user
* repr(): Required code that reproduces object; Generate output for developer

"Required code that reproduces object" is impractical for `Status`, mainly because there is not good way to concisely include the payload, but even just the `.message()` can potentially be very long. Therefore it is best to simply use the default `__repr__`, as is the case for the vast majority of types of significant complexity.

PiperOrigin-RevId: 476406024