Make `absl::Time` from-python conversion (almost) identical to a Google-internal implementation under absl/python.
Note that the unit tests are unchanged (in the entire Google codebase).

The remaining very subtle differences between `type_caster<absl::Time>::load()` and the corresponding from-python conversion under absl/python are related to the pybind11 two-pass overload resolution feature that does not exist in PyCLIF.

As a side-effect, this change removes long-obsolete remnants of Python 2 support.

PiperOrigin-RevId: 529457660